// src/app/models/movie.ts
export class MovieAll {
    status: string;

    movie_id: number;
    cinema_id: string;
    dateShow: string;
    listTime: string;

    id: number;
    titleEng: string;
    titleViet: string;
    deration: string;
   pubDay: string;
    direct: string;
    actor: string;
    image: string;
    description: string;
    cate: string;


    cinema: string;
     imdbRating: string;
     rating: number;
}
export class Movie {
    status: string;


    // movie
    id: number;
    titleEng: string;
    titleViet: string;
    deration: string;
    pubDay: string;
    direct: string;
    actor: string;
    image: string;
    description: string;
    cate: string;
    Director: string;
    Actors: string;
    Language: string;
    imdbVotes: string;


    // cinema
    cinema: string;
     imdbRating: string;
     rating: number;
}